Fixed OrderByGLA. Changed delimiter for print output. Meta-GLAs only call Finalize on multi GLAs.
OrderByGLA now performs properly. It is now less efficient than TopKGLA for
a single ranking attribute, but it is more flexible.

The delimiter was changed for the output of data in the Print waypoint from
comma (',') to the pipe ('|'). This makes it play better with TPCH data,
which sometimes contains commas in strings. The delimiter will be changed
to be configurable sometime in the future. The delimiter for column names
and types still needs to be changed.

The ToString method for HString was changed to surround the value in double
quotes. This allows for the output delimiter, whatever it may be, to be
safely contained within the string. In the future, it may be useful to be
able to define the surrounding character for HStrings.

Note: the ToString method currently does not check to ensure that there
are no double quotes within the string. This will be required in the
future for correctness (any double quotes found within the string will
need to be prefixed with a backslash to escape them).

The testing facility was modified in accordance to the delimiter and HString
changes.

The Meta-GLAs GroupByGLA, Multiplexer, and Repeater were modified to only call
finalize on 'multi' GLAs. 'single' GLAs do not require a finalize method,
and as such calling Finalize on them is an error. None of the Meta-GLAs
currently support fragmented GLAs as members, although it would be
possible to do so.

The TPCH query files were modified in accordance with the changes to the
delimiters and such.

TPC-H Query 2 now works properly. Query 1 and 6 still work.
Queries 3-5 still need to be modified to work properly and tested.
TPC-H Queries 7+ still need to be translated to Piggy.