Fixed problem with HString (Issue #5). Changed queries and a few GLAs to match.
HString now creates a local copy of its string if it is not in the dictionary
and it is created using the standard constructors or is copied from an
HString that is using local storage for its string.

NOTE: This means that when writing GLAs, if you want to input in the GLA,
YOU NEED TO USE THE COPY CONSTRUCTOR. Otherwise, you may lose information
in HStrings.

In the future, a copy operation will be added for all types so that there is
an alternative to using the copy constructor, as that may not be possible
in all cases.