Added GLA iteration to the system.
There are a lot of things that need to be improved in the future. The GLA
waypoint needs to be made aware of terminating query exits so that it can
properly receive states from other GLAs.

There is the initial framework in place for using states received from other
GLAs. Currently, the waypoint will not send the proper StartProducing
messages to any of these waypoints, but this isn't even supported in the
parser yet so it's not a huge deal.

This seems to work, but of course there could be bugs.