Further improvements to type checking, made most of base library external.
The parser now should (for the most part) automatically m4_include needed
files from libraries.

The code generation macros were changed slightly to allow the iterator
for a type to be specified via macros. This will allow for types to be more
easily added in the future. Also, types without an iterator defined are assumed
to not be serializable, and an m4 error will be thrown if a query attempts to
use a type in this manner (e.g., as an input or output to a waypoint).

Most of the base types were moved into the base library. The only types that
are still built into the system files are Bitstring and HString, due to how
tightly coupled with the system they are.