Fixed issue with the includes of typedefed GLAs. Added separator option for print.
A previous change to reduce the number of extra includes in a waypoint
unintentionally caused typedefed GLAs (which are not specified in a
waypoint) to have missing includes if one of their parameters was a
GLA. This has been corrected.

An option to specify the separator for printed output has been added to the
PRINT command. The separator can be any string.