Added State Passing to the system. Minor tweaks to base types.
It is now possible to pass states between waypoints without having to package
them into a chunk using the STATE datatype. The benefit of this is that these
states are received by a waypoint before it begins processing any chunks, and
references to the states are passed to the GLA in that waypoint upon creation.
This will allow for complex combinations of GLAs, and will be needed for
Generalized Filters and GISTs in the future.

A small example of this state-passing has been added. It consists of two GLAs:
Multimap, which creates a map from a key to a set of tuples that have that key,
and JoinGLA, which uses a Multimap to perform an asynchronous hash-based
equi-join. Using a GLA to perform the scan phase of the join is inefficient,
but this is just a proof of concept.