Improved compilation script.
The compilation script now determines the number of segments in the hash
table based on the total amount of memory in the system. It attempts to
set the number of segments such that the size of the hash table will be
within a certain range (currently 40% to 60%) of the total system memory.
It makes sure that the number of segments is a sum of only 1 or 2 powers
of 2.