TextLoader can now cache dropped chunks. Fixed issue with string serialization.
The TextLoader waypoing can now cache chunks, which means all of the data will
be processed by the system even if chunks are dropped. In addition, a
rudimentary throttling mechanism was added to the TextLoader to prevent dropped
chunks from being instantly resent. This throttling mechanism will likely be
replaced with a full-fledged congestion handling mechanism in the future.

Also, the TextLoader now has a new option, allowing the user to specify the
number of tuples per chunk that it reads.